Use a separate tag for tutorial code:
As tutorial code depends on both EJB and JLC (which,
in turn, depends on EJB messages and common library),
it cannot always be completed when _just_ EJB is released —
it may need a compatible JLC release also (if the previous
does not work with the new EJB). Hence, it have to
use a separate Git tag, that is set when both EJB and JLC
are released.

See the entries in the release procedure for details:
https://wiki.bf.local/display/EJB/Java+Binding+Release+Checklist+Template